fix: address code review — 4 Critical, 10 Important, 4 Minor
Critical
- C2 provider/rsclaw: drop blind retry on 502/504/429; only retry 503
  per docs §4.4. POST /turn is non-idempotent — retrying after a 502
  could double-append tool_results that already reached the worker.
- C3 config/env_resolution: SECRETREF detection now walks the JSON5
  Value tree (with sentinel substitution for unresolved placeholders)
  + comment-strip regex fallback, so {source:"env", id:"PATH"} inside
  a string literal or comment no longer leaks into .env.
- C4 agent/tools_session_archive: remove caller-controllable
  session_key arg; always use ctx.session_key. Closes a cross-session
  read hole when the runtime hosts multiple peers (Feishu/WeChat).
- C5 store/redb: post-sort archive_load by (generation, seq).
  Lexicographic key order placed gen10 between gen1 and gen2 so
  head:N / tail:N returned chronologically-wrong messages once a
  session had been /clear'd 10+ times. +regression test spanning 12
  generations.

Important
- I1 agent/tools_web: preserve raw markdown to its own artifact via
  preserve_raw_for_summarize() before maybe_summarize replaces text
  with a summary. attach_raw_artifact surfaces _tool_result_id so
  read_artifact still resolves the lossless source.
- I2 artifact/store: write-once via OpenOptions::create_new — id
  collisions error instead of silently truncating.
- I3 agent/runtime: squash_large_strings + format_tool_result before
  Phase-2b bus emit, so image data URLs / base64 don't dump raw
  payload into the streaming UI card.
- I5 config/env_resolution: SIGTERM → 500ms grace → SIGKILL so a
  wedged .zshrc no longer leaks the shell-fallback worker thread.
- I6 provider: widen TokenUsage fields to u64 across all four
  providers; drop every `as u32` truncation. extract_usage_count now
  returns u64.
- I7 cmd/env: delete unused redact() (panicked on multi-byte UTF-8)
  and the BTreeMap clippy workaround. Simplify `updated` to Vec<String>.
- I8 artifact/compact: build preview marker directly via new
  head_tail_with_marker helper. The old replacen of a generic marker
  silently no-op'd when content contained the same literal.
- UI-I1 sidebar / provider-card: centralize applyInstalledKey on
  sidebar-account-chip; provider-card downgraded to refresh-only.
  Removes triple read→merge→write race on rsclaw.json5.
- UI-I2 user-md-banner: narrow chat-store subscribe to the streaming
  true→false transition of the last assistant message — one disk
  read per turn instead of two.
- UI-I9 rsclaw-provider-card: mask API key input with type=password
  and a 👁/🙈 show/hide toggle.

Minor
- config/env_file: tolerate `export FOO=bar` lines pasted from a
  shell script. +regression test.
- memory-page: restructure filter row with outer justify-content
  space-between + inner controls wrapper; count label no longer
  drifts mid-row on narrow widths.
- toast: add neutral "info" type; sidebar ComingSoon now uses
  toast.info instead of the semantically-wrong toast.success.
- onboarding-rsclaw-card: document the home.tsx:269 invariant that
  justifies the dual applyInstalledKey path (SideBar is not mounted
  during onboarding, so no race with the chip).

Misc
- ui/package.json: add `refresh-gateway` npm script wrapping
  scripts/refresh-gateway.sh (dev workflow).

Held: C1 (gateway.a2aAuth → gateway.a2a deprecation alias) is left
for a separate follow-up.

+++ src/provider/gemini.rs
@@ -328,12 +328,12 @@ fn parse_event(data: &str) -> Option<StreamEvent> {
     // Finish reason.
     if candidate.get("finishReason").is_some() {
         let usage = v.get("usageMetadata").map(|u| TokenUsage {
-            input: u["promptTokenCount"].as_u64().unwrap_or(0) as u32,
-            output: u["candidatesTokenCount"].as_u64().unwrap_or(0) as u32,
+            input: u["promptTokenCount"].as_u64().unwrap_or(0),
+            output: u["candidatesTokenCount"].as_u64().unwrap_or(0),
             // Gemini reports cache reads via `cachedContentTokenCount`;
             // no separate creation counter (cache is implicit).
             cache_creation: 0,
-            cache_read: u["cachedContentTokenCount"].as_u64().unwrap_or(0) as u32,
+            cache_read: u["cachedContentTokenCount"].as_u64().unwrap_or(0),
         });
         return Some(StreamEvent::Done { usage });
     }
